requirements.txt
fastapi
uvicorn
python-multipart
tensorflow
pillow
numpy
python-dotenv
pytest
streamlit
